:recycle: Refactor tests; splitting to match naming of file under test.
